Fix error handling;
The previous version expected Bluebirds promisification to take
care of the handling of http error, which it obviously wont.
The new version uses request-promise for this although that
library as well as request itself has been deprecater. This
however was the easiest fix for the current and rather big
issue.